requirements.txt
Flask==3.1.2
paho_mqtt==2.1.0
